# customer-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing restaurant customers.",
        "title": "CJ Restaurant Customer-Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/customers": {
            "post": {
                "description": "Creates a new customer for a restaurant table",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Create a new customer",
                "parameters": [
                    {
                        "description": "Customer data",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateCustomerDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.Customer"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to create customer",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates an existing customer's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Update a customer",
                "parameters": [
                    {
                        "description": "Customer update data",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateCustomerDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to update customer",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/restaurant-tables/pending-delivery": {
            "get": {
                "description": "Returns a list of restaurant tables where orders are pending delivery",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "restaurant-tables"
                ],
                "summary": "Get tables with pending delivery",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.RestaurantTable"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get pending delivery tables",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/restaurant-tables/pending-payment": {
            "get": {
                "description": "Returns a list of restaurant tables where customers have not yet paid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "restaurant-tables"
                ],
                "summary": "Get tables with pending payment",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.RestaurantTable"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get pending payment tables",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.CreateCustomerDTO": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "integer"
                },
                "restaurant_table_id": {
                    "type": "integer"
                }
            }
        },
        "dtos.Customer": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "restaurant_table_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dtos.RestaurantTable": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dtos.UpdateCustomerDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}