# customer-service/internal/data/models/restaurant_table.go
package models

import "time"

// Database Model

type RestaurantTable struct {
	ID        int        `db:"id"`
	CreatedAt time.Time  `db:"created_at"`
	UpdatedAt time.Time  `db:"updated_at"`
	DeletedAt *time.Time `db:"deleted_at"`
	Status    string     `db:"status"`
}